Replace platform dependent aligned allocation by C++ operator new/delete

+++ examples/vectoradd/vectoradd.cpp
@@ -104,7 +104,7 @@ namespace usellama
                       << "s\n";
         }
 
-        return c.blob[0][0];
+        return (int)c.blob[0][0];
     }
 }
